Refactor puppeteer and e2e usage
