+++ package.json
@@ -22,18 +22,17 @@
     "build:electron:mac": "electron-builder build --mac --publish never",
     "build:electron:linux": "electron-builder build --linux --publish never",
     "test": "npm run test:app && npm run test:server",
-    "test:app": "vitest run --exclude \"**/pipelines.test.ts\"",
-    "test:app:covered": "vitest run --exclude \"**/e2e/*.test.ts\"",
+    "test:app": "vitest run --exclude \"**/e2e/*.test.ts\"",
     "test:tutorial": "vitest tutorial",
     "test:pipelines": "vitest pipelines",
     "test:progress": "vitest progress",
     "test:metadata": "vitest metadata",
-    "test:coverage:minimal": "npm run test:app:covered && npm run coverage:server",
     "test:server": "pytest src/pyflask/tests/ -s -vv",
     "wait5s": "node -e \"setTimeout(() => process.exit(0),5000)\"",
     "test:executable": "concurrently -n EXE,TEST --kill-others --success first \"node tests/testPyinstallerExecutable.js --port 3434 --forever\" \"npm run wait5s && pytest src/pyflask/tests/ -s --target http://localhost:3434\"",
     "test:coverage": "npm run coverage:app && npm run coverage:server",
     "coverage:app": "npm run test:app -- --coverage",
+    "coverage:tutorial": "npm run test:tutorial -- --coverage",
     "coverage:server": "pytest src/pyflask/tests/ -s --cov=pyflask --cov-report=xml",
     "deploy:win": "npm run build && npm run build:flask && npm run deploy:electron:win",
     "deploy:mac": "npm run build && npm run build:flask && npm run deploy:electron:mac",
