+++ tests/e2e/config.ts
@@ -4,7 +4,7 @@ import { homedir } from 'node:os'
 import { existsSync } from 'node:fs'
 
 import paths from "../../src/paths.config.json" assert { type: "json" };
-import { connect as connectToElectron } from '../puppeteer';
+import { connect as connectToElectron } from './puppeteer';
 
 // ------------------------------------------------------------------
 // ------------------------ Path Definitions ------------------------
