+++ tests/e2e/tutorial.test.ts
@@ -1,5 +1,5 @@
 import { beforeAll, describe, expect, test } from 'vitest'
-import { sleep } from '../puppeteer'
+import { sleep } from '../utils.js'
 
 import { existsSync } from 'node:fs'
 import { join } from 'node:path'
